# src/docs/openApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API RESTful Movilidad",
    "version": "1.0.0",
    "description": "Documentación interactiva de la API RESTful Movilidad"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "historyApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key requerida para los endpoints de historial"
      }
    },
    "schemas": {
...
      "RouteHistory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "userId": { "type": "integer", "example": 5 },
          "originLat": { "type": "number", "format": "float", "example": 20.6543 },
          "originLng": { "type": "number", "format": "float", "example": -103.3254 },
          "destinationLat": { "type": "number", "format": "float", "example": 20.6732 },
          "destinationLng": { "type": "number", "format": "float", "example": -103.3476 },
          "distanceM": { "type": "integer", "example": 3500 },
          "durationS": { "type": "integer", "example": 600 },
          "mode": { "type": "string", "example": "car" },
          "usedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-22T16:30:00Z"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "example": { "weather": "sunny", "notes": "Ruta rápida" }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-22T16:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-22T16:10:00Z"
          }
        }
      },
      "CreateRouteHistoryRequest": {
        "type": "object",
        "required": ["originLat", "originLng", "destinationLat", "destinationLng"],
        "properties": {
          "originLat": { "type": "number", "format": "float", "example": 20.6543 },
          "originLng": { "type": "number", "format": "float", "example": -103.3254 },
          "destinationLat": { "type": "number", "format": "float", "example": 20.6732 },
          "destinationLng": { "type": "number", "format": "float", "example": -103.3476 },
          "distanceM": { "type": "integer", "example": 3500 },
          "durationS": { "type": "integer", "example": 600 },
          "mode": { "type": "string", "example": "car" },
          "usedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-22T16:30:00Z"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "example": { "notes": "Ida a CUT" }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "userId inválido" },
          "details": {
            "type": "array",
            "items": { "type": "string" },
            "example": [
              "originLat inválido",
              "originLng inválido"
            ]
          }
        }
      }
    },
    "responses": {
...
      "ServerError": { "description": "Error interno del servidor" }
    },
    "parameters": {
      "UserIdParam": {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 },
        "description": "ID del usuario dueño del historial"
      },
      "HistoryIdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 },
        "description": "ID del registro de historial"
      }
    }
  },
  "paths": {
...
    "/history/{userId}": {
      "get": {
        "tags": ["History"],
        "summary": "Obtener historial de rutas de un usuario",
        "security": [{ "historyApiKey": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/UserIdParam" }
        ],
        "responses": {
          "200": {
            "description": "Lista de rutas del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Historial obtenido correctamente"
                    },
                    "history": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/RouteHistory" }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Sin registros para ese usuario",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["History"],
        "summary": "Crear un nuevo registro de historial de ruta para un usuario",
        "security": [{ "historyApiKey": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/UserIdParam" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateRouteHistoryRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ruta creada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Ruta creada correctamente"
                    },
                    "route": { "$ref": "#/components/schemas/RouteHistory" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Body inválido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/history/{userId}/{id}": {
      "put": {
        "tags": ["History"],
        "summary": "Actualizar un registro de historial",
        "security": [{ "historyApiKey": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/UserIdParam" },
          { "$ref": "#/components/parameters/HistoryIdParam" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "originLat": { "type": "number", "format": "float" },
                  "originLng": { "type": "number", "format": "float" },
                  "destinationLat": { "type": "number", "format": "float" },
                  "destinationLng": { "type": "number", "format": "float" },
                  "distanceM": { "type": "integer", "minimum": 0 },
                  "durationS": { "type": "integer", "minimum": 0 },
                  "mode": { "type": "string" },
                  "usedAt": { "type": "string", "format": "date-time" },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ruta actualizada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Ruta actualizada correctamente"
                    },
                    "route": { "$ref": "#/components/schemas/RouteHistory" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros o body inválidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Registro no encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "tags": ["History"],
        "summary": "Eliminar un registro de historial",
        "security": [{ "historyApiKey": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/UserIdParam" },
          { "$ref": "#/components/parameters/HistoryIdParam" }
        ],
        "responses": {
          "200": {
            "description": "Ruta eliminada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Ruta eliminada correctamente"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  }
}